See details
the mason package name for 'lua-language-server' is 'lua-language-server' and the lspconfig name
for it is 'lua_ls'. I don't think it specifies anywhere that the package
names must match, so I changed the way I specified names and added a
'mason_name' option.
Also made the way things are done a bit nicer. (At least I think so)